Move email notification method into plugin Fixes #2385
