+++ InvenTree/plugin/mixins/__init__.py
@@ -3,6 +3,7 @@ Utility class to enable simpler imports
 """
 
 from ..builtin.integration.mixins import APICallMixin, AppMixin, LabelPrintingMixin, SettingsMixin, EventMixin, ScheduleMixin, UrlsMixin, NavigationMixin
+from common.notifications import SingleNotificationMethod, BulkNotificationMethod
 
 from ..builtin.action.mixins import ActionMixin
 from ..builtin.barcode.mixins import BarcodeMixin
@@ -18,4 +19,6 @@ __all__ = [
     'UrlsMixin',
     'ActionMixin',
     'BarcodeMixin',
+    'SingleNotificationMethod',
+    'BulkNotificationMethod',
 ]
